refactor: 拆 _para.py 508 行 → 4 个 mixin 文件
把 5 个 analysis_s2d* 方法(s2d1/s2d2/s2d3/s2d5)挪到独立 mixin 文件,
通过多重继承拼到 dict_ 类上。

- _analysis_s2d1.py  (79 行) - 资产负债率趋势
- _analysis_s2d2.py  (92 行) - 短债/长债期限结构
- _analysis_s2d3.py (162 行) - 营收/净利润趋势
- _analysis_s2d5.py  (47 行) - 资产构成 TOP 5 (含 sort_asset)
- _para.py: 508 → 192 行 (含 6 个 dict 字面量 + 骨架)

5 个 mixin 跨文件无循环引用。dict_ 类签名/属性/方法名/返回 dict 形状
完全不变,main.py / start.py / webapp.py 不动。

smoke test 5/5 全过,docx 字节数与拆分前完全一致(5 套数据形态)。
